Divide Amount when submit Shared records

+++ UI/Interfaces/IForm.cs
@@ -6,7 +6,7 @@ namespace UI.Interfaces
 {
 	public interface IForm
 	{
-		int Amount { get; set; }
+		decimal Amount { get; set; }
 		void Submit();
 	}
 }
